aeon 0.4.0 and duplicate estimator removal
