+++ pyproject.toml
@@ -37,7 +37,7 @@ classifiers = [
     "Programming Language :: Python :: 3.11",
 ]
 dependencies = [
-    "aeon>=0.3.0,<0.4.0",
+    "aeon>=0.4.0,<0.5.0",
     "scikit-learn>=1.0.2,<=1.2.2",
     "tsml>=0.2.0,<0.3.0",
     "gpustat",
@@ -49,7 +49,6 @@ all_extras = [
     "aeon[all_extras,dl]",
     "tsml[extras]",
     "xgboost<=1.7.5",
-    "convst>=0.3.0,<=0.4.0; python_version < '3.11'",
     "torch>=1.13.1",
 ]
 unstable_extras = [
