+++ tsml_eval/estimators/__init__.py
@@ -1,4 +1,3 @@
-# -*- coding: utf-8 -*-
 """Estimators not present in tsml or aeon."""
 
 __all__ = [
@@ -7,8 +6,6 @@ __all__ = [
     "SklearnToTsmlRegressor",
 ]
 
-from tsml_eval.estimators.classification._sklearn_classifier import (
-    SklearnToTsmlClassifier,
-)
-from tsml_eval.estimators.clustering._sklearn_clusterer import SklearnToTsmlClusterer
-from tsml_eval.estimators.regression._sklearn_regressor import SklearnToTsmlRegressor
+from tsml_eval.estimators._sklearn_classifier import SklearnToTsmlClassifier
+from tsml_eval.estimators._sklearn_clusterer import SklearnToTsmlClusterer
+from tsml_eval.estimators._sklearn_regressor import SklearnToTsmlRegressor
