+++ tsml_eval/estimators/_sklearn_classifier.py
@@ -1,35 +1,35 @@
-# -*- coding: utf-8 -*-
-"""A tsml wrapper for sklearn regressors."""
+"""A tsml wrapper for sklearn classifiers."""
 
-__author__ = ["DGuijoRubio", "MatthewMiddlehurst"]
-__all__ = ["SklearnToTsmlRegressor"]
+__author__ = ["MatthewMiddlehurst"]
+__all__ = ["SklearnToTsmlClassifier"]
 
 import numpy as np
-from sklearn.base import RegressorMixin
+from sklearn.base import ClassifierMixin
+from sklearn.utils.multiclass import check_classification_targets
 from sklearn.utils.validation import check_is_fitted
 from tsml.base import BaseTimeSeriesEstimator, _clone_estimator
 
 
-class SklearnToTsmlRegressor(RegressorMixin, BaseTimeSeriesEstimator):
+class SklearnToTsmlClassifier(ClassifierMixin, BaseTimeSeriesEstimator):
     """Wrapper for sklearn estimators."""
 
     def __init__(
         self,
-        regressor=None,
+        classifier=None,
         pad_unequal=False,
         concatenate_channels=False,
         random_state=None,
     ):
-        self.regressor = regressor
+        self.classifier = classifier
         self.pad_unequal = pad_unequal
         self.concatenate_channels = concatenate_channels
         self.random_state = random_state
 
-        super(SklearnToTsmlRegressor, self).__init__()
+        super(SklearnToTsmlClassifier, self).__init__()
 
     def fit(self, X, y):
-        if self.regressor is None:
-            raise ValueError("Regressor not set")
+        if self.classifier is None:
+            raise ValueError("Classifier not set")
 
         X, y = self._validate_data(X=X, y=y)
         X = self._convert_X(
@@ -38,8 +38,11 @@ class SklearnToTsmlRegressor(RegressorMixin, BaseTimeSeriesEstimator):
             concatenate_channels=self.concatenate_channels,
         )
 
-        self._regressor = _clone_estimator(self.regressor, self.random_state)
-        self._regressor.fit(X, y)
+        check_classification_targets(y)
+        self.classes_ = np.unique(y)
+
+        self._classifier = _clone_estimator(self.classifier, self.random_state)
+        self._classifier.fit(X, y)
 
         return self
 
@@ -53,7 +56,19 @@ class SklearnToTsmlRegressor(RegressorMixin, BaseTimeSeriesEstimator):
             concatenate_channels=self.concatenate_channels,
         )
 
-        return self._regressor.predict(X)
+        return self._classifier.predict(X)
+
+    def predict_proba(self, X) -> np.ndarray:
+        check_is_fitted(self)
+
+        X = self._validate_data(X=X, reset=False)
+        X = self._convert_X(
+            X,
+            pad_unequal=self.pad_unequal,
+            concatenate_channels=self.concatenate_channels,
+        )
+
+        return self._classifier.predict_proba(X)
 
     def _more_tags(self):
         return {
