+++ tsml_eval/estimators/_sklearn_clusterer.py
@@ -1,49 +1,46 @@
-# -*- coding: utf-8 -*-
-"""A tsml wrapper for sklearn classifiers."""
+"""A tsml wrapper for sklearn clusterers."""
 
 __author__ = ["MatthewMiddlehurst"]
-__all__ = ["SklearnToTsmlClassifier"]
+__all__ = ["SklearnToTsmlClusterer"]
 
 import numpy as np
-from sklearn.base import ClassifierMixin
-from sklearn.utils.multiclass import check_classification_targets
+from sklearn.base import ClusterMixin
 from sklearn.utils.validation import check_is_fitted
 from tsml.base import BaseTimeSeriesEstimator, _clone_estimator
 
 
-class SklearnToTsmlClassifier(ClassifierMixin, BaseTimeSeriesEstimator):
+class SklearnToTsmlClusterer(ClusterMixin, BaseTimeSeriesEstimator):
     """Wrapper for sklearn estimators."""
 
     def __init__(
         self,
-        classifier=None,
+        clusterer=None,
         pad_unequal=False,
         concatenate_channels=False,
         random_state=None,
     ):
-        self.classifier = classifier
+        self.clusterer = clusterer
         self.pad_unequal = pad_unequal
         self.concatenate_channels = concatenate_channels
         self.random_state = random_state
 
-        super(SklearnToTsmlClassifier, self).__init__()
+        super(SklearnToTsmlClusterer, self).__init__()
 
-    def fit(self, X, y):
-        if self.classifier is None:
-            raise ValueError("Classifier not set")
+    def fit(self, X, y=None):
+        if self.clusterer is None:
+            raise ValueError("Clusterer not set")
 
-        X, y = self._validate_data(X=X, y=y)
+        X = self._validate_data(X=X)
         X = self._convert_X(
             X,
             pad_unequal=self.pad_unequal,
             concatenate_channels=self.concatenate_channels,
         )
 
-        check_classification_targets(y)
-        self.classes_ = np.unique(y)
+        self._clusterer = _clone_estimator(self.clusterer, self.random_state)
+        self._clusterer.fit(X, y)
 
-        self._classifier = _clone_estimator(self.classifier, self.random_state)
-        self._classifier.fit(X, y)
+        self.labels_ = self._clusterer.labels_
 
         return self
 
@@ -57,19 +54,7 @@ class SklearnToTsmlClassifier(ClassifierMixin, BaseTimeSeriesEstimator):
             concatenate_channels=self.concatenate_channels,
         )
 
-        return self._classifier.predict(X)
-
-    def predict_proba(self, X) -> np.ndarray:
-        check_is_fitted(self)
-
-        X = self._validate_data(X=X, reset=False)
-        X = self._convert_X(
-            X,
-            pad_unequal=self.pad_unequal,
-            concatenate_channels=self.concatenate_channels,
-        )
-
-        return self._classifier.predict_proba(X)
+        return self._clusterer.predict(X)
 
     def _more_tags(self):
         return {
