+++ tsml_eval/estimators/_sklearn_regressor.py
@@ -1,47 +1,44 @@
-# -*- coding: utf-8 -*-
-"""A tsml wrapper for sklearn clusterers."""
+"""A tsml wrapper for sklearn regressors."""
 
-__author__ = ["MatthewMiddlehurst"]
-__all__ = ["SklearnToTsmlClusterer"]
+__author__ = ["DGuijoRubio", "MatthewMiddlehurst"]
+__all__ = ["SklearnToTsmlRegressor"]
 
 import numpy as np
-from sklearn.base import ClusterMixin
+from sklearn.base import RegressorMixin
 from sklearn.utils.validation import check_is_fitted
 from tsml.base import BaseTimeSeriesEstimator, _clone_estimator
 
 
-class SklearnToTsmlClusterer(ClusterMixin, BaseTimeSeriesEstimator):
+class SklearnToTsmlRegressor(RegressorMixin, BaseTimeSeriesEstimator):
     """Wrapper for sklearn estimators."""
 
     def __init__(
         self,
-        clusterer=None,
+        regressor=None,
         pad_unequal=False,
         concatenate_channels=False,
         random_state=None,
     ):
-        self.clusterer = clusterer
+        self.regressor = regressor
         self.pad_unequal = pad_unequal
         self.concatenate_channels = concatenate_channels
         self.random_state = random_state
 
-        super(SklearnToTsmlClusterer, self).__init__()
+        super(SklearnToTsmlRegressor, self).__init__()
 
-    def fit(self, X, y=None):
-        if self.clusterer is None:
-            raise ValueError("Clusterer not set")
+    def fit(self, X, y):
+        if self.regressor is None:
+            raise ValueError("Regressor not set")
 
-        X = self._validate_data(X=X)
+        X, y = self._validate_data(X=X, y=y)
         X = self._convert_X(
             X,
             pad_unequal=self.pad_unequal,
             concatenate_channels=self.concatenate_channels,
         )
 
-        self._clusterer = _clone_estimator(self.clusterer, self.random_state)
-        self._clusterer.fit(X, y)
-
-        self.labels_ = self._clusterer.labels_
+        self._regressor = _clone_estimator(self.regressor, self.random_state)
+        self._regressor.fit(X, y)
 
         return self
 
@@ -55,7 +52,7 @@ class SklearnToTsmlClusterer(ClusterMixin, BaseTimeSeriesEstimator):
             concatenate_channels=self.concatenate_channels,
         )
 
-        return self._clusterer.predict(X)
+        return self._regressor.predict(X)
 
     def _more_tags(self):
         return {
