--- tsml_eval/estimators/classification/feature_based/__init__.py
@@ -1,6 +1,0 @@
-# -*- coding: utf-8 -*-
-__all__ = ["FreshPRINCEClassifier"]
-
-from tsml_eval.estimators.classification.feature_based._fresh_prince import (
-    FreshPRINCEClassifier,
-)
