--- tsml_eval/estimators/classification/hybrid/__init__.py
@@ -1,5 +1,0 @@
-"""Hyybrid classifiers."""
-
-__all__ = ["HIVECOTEV1"]
-
-from tsml_eval.estimators.classification.hybrid.hivecote_v1 import HIVECOTEV1
