--- tsml_eval/estimators/classification/shapelet_based/__init__.py
@@ -1,9 +1,0 @@
-# -*- coding: utf-8 -*-
-"""Shapelet-based classification algorithms."""
-
-__all__ = ["RDST", "RDSTEnsemble", "ShapeletTransformClassifier"]
-
-from tsml_eval.estimators.classification.shapelet_based.rdst import RDST, RDSTEnsemble
-from tsml_eval.estimators.classification.shapelet_based.stc import (
-    ShapeletTransformClassifier,
-)
