--- tsml_eval/estimators/classification/sklearn/__init__.py
@@ -1,4 +1,0 @@
-# -*- coding: utf-8 -*-
-__all__ = ["RotationForestClassifier"]
-
-from tsml_eval.estimators.classification.sklearn.rotf import RotationForestClassifier
