--- tsml_eval/estimators/clustering/__init__.py
@@ -1,1 +1,0 @@
-"""Clustering estimators."""
