--- tsml_eval/estimators/networks/__init__.py
@@ -1,2 +1,0 @@
-# -*- coding: utf-8 -*-
-"""Networks init."""
